initial commit
Started a project and begin to implement a pop-up that will show on
first load and get the graduation date from the user.

Currently, the XAML toolkit for Windows 8 does not include a DatePicker.
There's a company named Telerik that offers their own library full of
XAML controls and includes a DatePicker, but it's $200. The open-source
Callisto library for Windows 8 advertises that they will have one, but
it is not yet completed. I will probably need to implement my own date
picker before moving on, I might ghetto hook it up with 3 drop down
menus or something.