MVP complete!
Had a major code session today and achieved a lot of the major
functionality. Upon installing the app users will be prompted for their
graduation date which will be saved locally. I added a top app bar and a
button on there for users to change their graduation date. The countdown
code was the same as was used in my Christmas Countdown so that was easy
to complete.

TODO:
- Artwork
- Live tile
- Other cool settings/ customization options?